requirements.txt
pydub
SpeechRecognition
openai-whisper
torch
